@@ -1,3 +1,3 @@
-﻿@{
-    Layout = "/Views/Shared/_Layout.cshtml";
+@{
+  Layout = "/Views/Shared/_Layout.cshtml";
 }
