@@ -1,4 +1,7 @@
-﻿@using BestClipOfTheWeek
+@using Microsoft.AspNetCore.Identity
+@using BestClipOfTheWeek
 @using BestClipOfTheWeek.Models
 @using BestClipOfTheWeek.Extensions
+@using BestClipOfTheWeek.Data
+@namespace BestClipOfTheWeek.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
